Added testing on number to letter bounds Making use of Byte type and using Long for inputs

+++ phone-number-to-word-converter/src/main/java/com/mark/phoneword/DefaultNumberToLettersConverter.java
@@ -11,18 +11,18 @@ import java.util.stream.Collectors;
  */
 class DefaultNumberToLettersConverter extends NumberToLettersConverter {
 
-    private final static Map<Integer, Set<Character>> DEFAULT_CONVERSION_MAP;
+    private final static Map<Byte, Set<Character>> DEFAULT_CONVERSION_MAP;
 
     static {
         DEFAULT_CONVERSION_MAP = new HashMap<>();
-        DEFAULT_CONVERSION_MAP.put(2, new HashSet<>(Arrays.asList('a', 'b', 'c')));
-        DEFAULT_CONVERSION_MAP.put(3, new HashSet<>(Arrays.asList('d', 'e', 'f')));
-        DEFAULT_CONVERSION_MAP.put(4, new HashSet<>(Arrays.asList('g', 'h', 'i')));
-        DEFAULT_CONVERSION_MAP.put(5, new HashSet<>(Arrays.asList('j', 'k', 'l')));
-        DEFAULT_CONVERSION_MAP.put(6, new HashSet<>(Arrays.asList('m', 'n', 'o')));
-        DEFAULT_CONVERSION_MAP.put(7, new HashSet<>(Arrays.asList('p', 'q', 'r', 's')));
-        DEFAULT_CONVERSION_MAP.put(8, new HashSet<>(Arrays.asList('t', 'u', 'v')));
-        DEFAULT_CONVERSION_MAP.put(9, new HashSet<>(Arrays.asList('w', 'x', 'y', 'z')));
+        DEFAULT_CONVERSION_MAP.put((byte)2, new HashSet<>(Arrays.asList('a', 'b', 'c')));
+        DEFAULT_CONVERSION_MAP.put((byte)3, new HashSet<>(Arrays.asList('d', 'e', 'f')));
+        DEFAULT_CONVERSION_MAP.put((byte)4, new HashSet<>(Arrays.asList('g', 'h', 'i')));
+        DEFAULT_CONVERSION_MAP.put((byte)5, new HashSet<>(Arrays.asList('j', 'k', 'l')));
+        DEFAULT_CONVERSION_MAP.put((byte)6, new HashSet<>(Arrays.asList('m', 'n', 'o')));
+        DEFAULT_CONVERSION_MAP.put((byte)7, new HashSet<>(Arrays.asList('p', 'q', 'r', 's')));
+        DEFAULT_CONVERSION_MAP.put((byte)8, new HashSet<>(Arrays.asList('t', 'u', 'v')));
+        DEFAULT_CONVERSION_MAP.put((byte)9, new HashSet<>(Arrays.asList('w', 'x', 'y', 'z')));
     }
 
     DefaultNumberToLettersConverter() {
